+++ migrations/m171015_214532_create_table_group_role.php
@@ -13,8 +13,8 @@ class m171015_214532_create_table_group_role extends Migration
 
         $this->createTable('{{%group_role}}', [
             'id' => $this->integer(11)->notNull()->append('AUTO_INCREMENT PRIMARY KEY'),
-            'id_group' => $this->integer(10)->unsigned()->notNull(),
-            'id_rol' => $this->integer(10)->unsigned()->notNull(),
+            'id_group' => $this->integer(11)->unsigned()->notNull(),
+            'id_rol' => $this->integer(11)->unsigned()->notNull(),
             'date_modified' => $this->timestamp()->notNull()->defaultExpression('CURRENT_TIMESTAMP'),
             'date_created' => $this->timestamp()->notNull()->defaultExpression('CURRENT_TIMESTAMP'),
         ], $tableOptions);
